DocDefinitions e buffer dos chunks separados entre o reportService e reportController
